# docs/swagger.yaml
basePath: /api/v1
definitions:
  CreatePasteBody:
    description: Тело запроса для создания пасты.
    properties:
      expires:
        description: Время, через которое паста становится не доступной
        enum:
        - 30m
        - 1h
        - 168h
        - 5040h
        example: 30m
        type: string
      format:
        description: Формат текста
        enum:
        - json
        - yaml
        - toml
        example: plaintext
        type: string
      password:
        description: Пароль для получения доступа к пасте
        example: password for security
        maxLength: 255
        type: string
      text:
        description: Текст
        example: Some very secret text
        type: string
      title:
        description: Название
        example: The private paste
        maxLength: 255
        type: string
    required:
    - format
    - text
    type: object
  CreateTokenRequest:
    description: Payload for creating a new user if not exists and get access token.
    properties:
      code:
        description: Github oauth2 code
        type: string
    type: object
  PasteResponse:
    description: Тело ответа на создание пасты.
    properties:
      created_at:
        description: Дата создания
        type: string
      expires_at:
        description: Время, через которое паста становится не доступной
        type: string
      format:
        description: Формат текста
        type: string
      hash:
        description: Уникальный идентификатор
        type: string
      text:
        description: Текст
        type: string
      title:
        description: Название
        type: string
    type: object
  UnlockPasteBody:
    description: Тело запроса для разблокировки пасты.
    properties:
      password:
        description: Пароль
        example: hello
        maxLength: 255
        type: string
    required:
    - password
    type: object
  UserInfo:
    description: Payload for getting user info.
    properties:
      avatar_url:
        type: string
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: Implementation pastebin API
  title: Pastebin API
  version: "1.0"
paths:
  /:
    post:
      consumes:
      - application/json
      parameters:
      - description: Уникальный код, сгенерированный OAuth2 приложением
        in: body
        name: code
        required: true
        schema:
          $ref: '#/definitions/CreateTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  user:
                    $ref: '#/definitions/UserInfo'
                type: object
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
            type: object
      summary: Регистрация нового пользователя с помощью Github OAuth 2.0
      tags:
      - auth
  /pastes:
    post:
      consumes:
      - application/json
      parameters:
      - description: Паста
        in: body
        name: paste
        required: true
        schema:
          $ref: '#/definitions/CreatePasteBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  paste:
                    $ref: '#/definitions/PasteResponse'
                  url:
                    type: string
                type: object
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            properties:
              errors:
                properties:
                  field:
                    type: string
                  message:
                    type: string
                type: object
              message:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
            type: object
      summary: Создание нововой пасты
      tags:
      - pastes
  /pastes/{hash}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Хеш пасты
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Удаление пасты по хешу
      tags:
      - pastes
    get:
      description: |-
        Получение посты по хешу.
        Если паста защищена паролем, то нужно обратиться к `/pastes/{hash}/unlock`, чтобы получить доступ к ней.
      parameters:
      - description: Хеш пасты
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  paste:
                    $ref: '#/definitions/PasteResponse'
                type: object
              message:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Получениие пасты.
      tags:
      - pastes
  /pastes/{hash}/unlock:
    post:
      consumes:
      - application/json
      parameters:
      - description: Хеш пасты
        in: path
        name: hash
        required: true
        type: string
      - description: Пароль
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/UnlockPasteBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  paste:
                    $ref: '#/definitions/PasteResponse'
                type: object
              message:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Получение доступа к пасте с паролем.
      tags:
      - pastes
  /token:
    post:
      consumes:
      - application/json
      parameters:
      - description: Уникальный код, сгенерированный OAuth2 приложением
        in: body
        name: code
        required: true
        schema:
          $ref: '#/definitions/CreateTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
            type: object
      summary: Получения авторизационных данных
      tags:
      - auth
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"